+++ packages/electron-tools/src/config/read-config.ts
@@ -2,8 +2,7 @@
  * read user config
  */
 
-import { cosmiconfigSync, OptionsSync, defaultLoaders } from 'cosmiconfig';
-import { TypeScriptLoader } from "cosmiconfig-typescript-loader";
+import { cosmiconfigSync } from 'cosmiconfig';
 import { UserConfig } from './interface';
 
 export const DEFAULT_CONFIG_NAME = 'electron';
@@ -14,44 +13,23 @@ export const readConfig = (projectPath: string) => {
   if (defaultElectronConfig) {
     return defaultElectronConfig as UserConfig;
   }
-  const jupiterElectronTsConfig = doReadConfig(projectPath, MODERN_CONFIG_NAME, true);
-  if (jupiterElectronTsConfig) {
-    return (jupiterElectronTsConfig?.electron || {}) as UserConfig;
-  }
   const jupiterElectronConfig = doReadConfig(projectPath, MODERN_CONFIG_NAME);
   return (jupiterElectronConfig?.electron || {}) as UserConfig;
 };
 
-const doReadConfig = (projectPath: string, moduleName: string, isTs?: boolean) => {
-  let defaultCosmiconfig: OptionsSync = {
-    stopDir: projectPath,
-    searchPlaces: [
-      `.${moduleName}rc`,
-      `.${moduleName}rc.json`,
-      `.${moduleName}rc.yaml`,
-      `.${moduleName}rc.yml`,
-      `.${moduleName}rc.js`,
-      `${moduleName}.config.js`,
-    ]
-  }
-
-  if (isTs) {
-    defaultCosmiconfig.loaders = {
-      ...defaultLoaders,
-      ".ts": TypeScriptLoader({
-        transpileOnly: true,
-        emit: false,
-        compilerOptions: {
-          sourceMap: false,
-        }
-      })
-    }
-  }
-  const explorer = cosmiconfigSync(moduleName, defaultCosmiconfig);
-  if (isTs) {
-    const loaded = explorer.load(`${moduleName}.config.ts`);
-    return loaded?.config || null;
-  }
-  const searched = explorer.search(projectPath);
+const doReadConfig = (projectPath: string, moduleName: string) => {
+  const explorer = (_moduleName: string, _projectPath: string) =>
+    cosmiconfigSync(moduleName, {
+      stopDir: projectPath,
+      searchPlaces: [
+        `.${moduleName}rc`,
+        `.${moduleName}rc.json`,
+        `.${moduleName}rc.yaml`,
+        `.${moduleName}rc.yml`,
+        `.${moduleName}rc.js`,
+        `${moduleName}.config.js`,
+      ],
+    });
+  const searched = explorer(moduleName, projectPath).search(projectPath);
   return searched?.config || null;
 };